fix(canonical): make chain sequences portable
Use an int64 canonical sequence lane so the shared 0..2^53-1 identity domain remains representable on 32-bit Go. Propagate the type through journal, gauntlet, runtime, and proto callers, with checked narrowing at platform-sized cursor and index boundaries.

Evidence: GOARCH=386 canonical and journal boundary tests; native Go vet/test and journal race; proto Go/TS gates; root Bun gates; js/wasm compile and runnable wall; shared CG1 gate; frozen fixtures byte-identical.

+++ go/canonical/conformance_test.go
@@ -143,7 +143,7 @@ func TestChainFoldConsistency(t *testing.T) {
 	}
 	prev := canonical.Genesis
 	for i, p := range payloads {
-		want := mustEntryDigest(t, canonical.ChainEntry{Seq: i, Prev: prev, Payload: p})
+		want := mustEntryDigest(t, canonical.ChainEntry{Seq: int64(i), Prev: prev, Payload: p})
 		if digests[i] != want {
 			t.Fatalf("entry %d (payload %q): got %s want %s", i, p, digests[i], want)
 		}
